refactor: keep explosion exposure with explosions
